+++ packages/framework/src/Controller/Admin/TransportController.php
@@ -9,7 +9,6 @@ use Shopsys\FrameworkBundle\Form\Admin\Transport\TransportEditFormType;
 use Shopsys\FrameworkBundle\Model\AdminNavigation\Breadcrumb;
 use Shopsys\FrameworkBundle\Model\AdminNavigation\MenuItem;
 use Shopsys\FrameworkBundle\Model\Pricing\Currency\CurrencyFacade;
-use Shopsys\FrameworkBundle\Model\Transport\Detail\TransportDetailFactory;
 use Shopsys\FrameworkBundle\Model\Transport\Grid\TransportGridFactory;
 use Shopsys\FrameworkBundle\Model\Transport\TransportDataFactory;
 use Shopsys\FrameworkBundle\Model\Transport\TransportFacade;
@@ -26,11 +25,6 @@ class TransportController extends AdminBaseController
      * @var \Shopsys\FrameworkBundle\Model\Pricing\Currency\CurrencyFacade
      */
     private $currencyFacade;
-
-    /**
-     * @var \Shopsys\FrameworkBundle\Model\Transport\Detail\TransportDetailFactory
-     */
-    private $transportDetailFactory;
 
     /**
      * @var \Shopsys\FrameworkBundle\Model\Transport\Grid\TransportGridFactory
@@ -52,14 +46,12 @@ class TransportController extends AdminBaseController
         TransportGridFactory $transportGridFactory,
         TransportDataFactory $transportDataFactory,
         CurrencyFacade $currencyFacade,
-        TransportDetailFactory $transportDetailFactory,
         Breadcrumb $breadcrumb
     ) {
         $this->transportFacade = $transportFacade;
         $this->transportGridFactory = $transportGridFactory;
         $this->transportDataFactory = $transportDataFactory;
         $this->currencyFacade = $currencyFacade;
-        $this->transportDetailFactory = $transportDetailFactory;
         $this->breadcrumb = $breadcrumb;
     }
 
@@ -72,7 +64,7 @@ class TransportController extends AdminBaseController
         $transportData = $this->transportDataFactory->createDefault();
 
         $form = $this->createForm(TransportEditFormType::class, $transportData, [
-            'transport_detail' => null,
+            'transport' => null,
         ]);
         $form->handleRequest($request);
 
@@ -107,11 +99,10 @@ class TransportController extends AdminBaseController
     public function editAction(Request $request, $id)
     {
         $transport = $this->transportFacade->getById($id);
-        $transportDetail = $this->transportDetailFactory->createDetailForTransportWithIndependentPrices($transport);
         $transportData = $this->transportDataFactory->createFromTransport($transport);
 
         $form = $this->createForm(TransportEditFormType::class, $transportData, [
-            'transport_detail' => $transportDetail,
+            'transport' => $transport,
         ]);
         $form->handleRequest($request);
 
@@ -136,7 +127,7 @@ class TransportController extends AdminBaseController
 
         return $this->render('@ShopsysFramework/Admin/Content/Transport/edit.html.twig', [
             'form' => $form->createView(),
-            'transportDetail' => $transportDetail,
+            'transport' => $transport,
             'currencies' => $this->currencyFacade->getAllIndexedById(),
         ]);
     }