+++ packages/framework/src/Model/Transport/Grid/TransportGridFactory.php
@@ -7,8 +7,8 @@ use Shopsys\FrameworkBundle\Component\Grid\GridFactory;
 use Shopsys\FrameworkBundle\Component\Grid\GridFactoryInterface;
 use Shopsys\FrameworkBundle\Component\Grid\QueryBuilderWithRowManipulatorDataSource;
 use Shopsys\FrameworkBundle\Model\Localization\Localization;
-use Shopsys\FrameworkBundle\Model\Transport\Detail\TransportDetailFactory;
 use Shopsys\FrameworkBundle\Model\Transport\Transport;
+use Shopsys\FrameworkBundle\Model\Transport\TransportFacade;
 use Shopsys\FrameworkBundle\Model\Transport\TransportRepository;
 
 class TransportGridFactory implements GridFactoryInterface
@@ -26,25 +26,25 @@ class TransportGridFactory implements GridFactoryInterface
     private $transportRepository;
 
     /**
-     * @var \Shopsys\FrameworkBundle\Model\Transport\Detail\TransportDetailFactory
-     */
-    private $transportDetailFactory;
-
-    /**
      * @var \Shopsys\FrameworkBundle\Model\Localization\Localization
      */
     private $localization;
 
+    /**
+     * @var \Shopsys\FrameworkBundle\Model\Transport\TransportFacade
+     */
+    private $transportFacade;
+
     public function __construct(
         GridFactory $gridFactory,
         TransportRepository $transportRepository,
-        TransportDetailFactory $transportDetailFactory,
-        Localization $localization
+        Localization $localization,
+        TransportFacade $transportFacade
     ) {
         $this->gridFactory = $gridFactory;
         $this->transportRepository = $transportRepository;
-        $this->transportDetailFactory = $transportDetailFactory;
         $this->localization = $localization;
+        $this->transportFacade = $transportFacade;
     }
 
     /**
@@ -61,7 +61,7 @@ class TransportGridFactory implements GridFactoryInterface
             't.id',
             function ($row) {
                 $transport = $this->transportRepository->findById($row['t']['id']);
-                $row['transportDetail'] = $this->transportDetailFactory->createDetailForTransportWithIndependentPrices($transport);
+                $row['displayPrice'] = $this->getDisplayPrice($transport);
                 return $row;
             }
         );
@@ -70,18 +70,26 @@ class TransportGridFactory implements GridFactoryInterface
         $grid->enableDragAndDrop(Transport::class);
 
         $grid->addColumn('name', 'tt.name', t('Name'));
-        $grid->addColumn('price', 'transportDetail', t('Price'));
+        $grid->addColumn('price', 'displayPrice', t('Price'));
 
         $grid->setActionColumnClassAttribute('table-col table-col-10');
         $grid->addEditActionColumn('admin_transport_edit', ['id' => 't.id']);
         $grid->addDeleteActionColumn('admin_transport_delete', ['id' => 't.id'])
             ->setConfirmMessage(t('Do you really want to remove this shipping?'));
 
-        $grid->setTheme(
-            '@ShopsysFramework/Admin/Content/Transport/listGrid.html.twig',
-            ['currencyIdForList' => self::CURRENCY_ID_FOR_LIST]
-        );
+        $grid->setTheme('@ShopsysFramework/Admin/Content/Transport/listGrid.html.twig');
 
         return $grid;
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Transport\Transport $transport
+     * @return string
+     */
+    private function getDisplayPrice(Transport $transport)
+    {
+        $transportBasePricesIndexedByCurrencyId = $this->transportFacade->getIndependentBasePricesIndexedByCurrencyId($transport);
+
+        return $transportBasePricesIndexedByCurrencyId[self::CURRENCY_ID_FOR_LIST]->getPriceWithVat();
     }
 }